% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{olapAddParameters}
\alias{olapAddParameters}
\alias{olapAddParameters,OlapCommand,OlapParameter-method}
\alias{olapAddParameters,OlapCommand,list-method}
\alias{olapAddParameters,OlapParameterCollection,OlapParameter-method}
\title{Add parameter(s) to command}
\usage{
olapAddParameters(x, params, ...)

\S4method{olapAddParameters}{OlapParameterCollection,OlapParameter}(x, params,
  ...)

\S4method{olapAddParameters}{OlapCommand,OlapParameter}(x, params, ...)

\S4method{olapAddParameters}{OlapCommand,list}(x, params, ...)
}
\description{
Add parameter(s) to command
}
\section{Methods (by class)}{
\itemize{
\item \code{x = OlapParameterCollection,params = OlapParameter}: 

\item \code{x = OlapCommand,params = OlapParameter}: 

\item \code{x = OlapCommand,params = list}: 
}}
\examples{
conn <- olapConnection("Data Source=localhost;")
open(conn)
cmd <- olapCommand("SELECT StrToSet(@Hierarchy) ON 0 FROM [Table] WHERE [Some].[Measure]", conn)
param <- olapParameter("Hierarchy", "[Some].[Hierarchy].Members")
olapAddParameters(param, cmd)
df <- as.data.frame(cmd)
}

