% Generated by roxygen2 (4.0.2): do not edit by hand
\name{olapCommand}
\alias{olapCommand}
\title{Create a new command}
\usage{
olapCommand(cmdText = NULL, conn = NULL)
}
\arguments{
\item{conn}{An \code{OlapConnection} object.}

\item{cmd}{A valid MDX command, either as a length one character vector or a reference to a file.}
}
\value{
An \code{OlapCommand} object.
}
\description{
Create a new command
}
\examples{
conn <- olapConnection("Data Source=localhost;")
open(conn)
cmd <- olapCommand("SELECT [Some].[Hierarchy].Members ON 0 FROM [Table] WHERE [Some].[Measure]", conn)
df <- as.data.frame(cmd)
}

