% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OlapConnection.R
\docType{methods}
\name{olapConnection}
\alias{olapConnection}
\alias{olapConnection,OlapConnection-method}
\alias{olapConnection,character-method}
\title{Create a connection}
\usage{
olapConnection(x = NULL)

\S4method{olapConnection}{character}(x = NULL)

\S4method{olapConnection}{OlapConnection}(x = NULL)
}
\value{
An \code{OlapConnection} object.
}
\description{
Create a connection
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Connect using a connection string. See \url{http://msdn.microsoft.com/en-us/library/dn140245.aspx}

\item \code{OlapConnection}: Connect using a template \code{\link{OlapConnection}}
}}
\examples{
conn <- olapConnection()
}

