% Generated by roxygen2 (4.0.2): do not edit by hand
\name{olapExecute}
\alias{olapExecute}
\title{Execute a command}
\usage{
olapExecute(cmd)
}
\arguments{
\item{cmd}{A command of class \code{OlapCommand}}
}
\value{
Boolean TRUE if executed successfully.
}
\description{
Execute a command
}
\examples{
parser <- olapExecute(cmd)
}

