% Generated by roxygen2 (4.0.2): do not edit by hand
\name{olapParameter}
\alias{olapParameter}
\title{Create a new parameter}
\usage{
olapParameter(name = NULL, value = NULL)
}
\arguments{
\item{name}{Parameter name}

\item{value}{Parameter value}
}
\value{
An \code{OlapParameter} object.
}
\description{
Create a new parameter
}
\examples{
conn <- olapConnection("Data Source=localhost;")
open(conn)
cmd <- olapCommand("SELECT StrToSet(@Hierarchy) ON 0 FROM [Table] WHERE [Some].[Measure]", conn)
param <- olapParameter("Hierarchy", "[Some].[Hierarchy].Members")
olapAddParameters(param, cmd)
df <- as.data.frame(cmd)
}

