% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{open,OlapConnection-method}
\alias{open,OlapConnection-method}
\title{Open a connection}
\usage{
\S4method{open}{OlapConnection}(con, newSession = TRUE, quietly = TRUE, ...)
}
\arguments{
\item{An}{\code{OlapConnection} object}
}
\description{
Open a connection
}
\examples{
open(conn)
}

