% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OlapConnection.R
\docType{methods}
\name{close,OlapConnection-method}
\alias{close,OlapConnection-method}
\title{Close a connection}
\usage{
\S4method{close}{OlapConnection}(con, endSession = TRUE, quietly = TRUE,
  ...)
}
\arguments{
\item{An}{\code{OlapConnection} object}
}
\description{
Close a connection
}
\examples{
close(conn)
}

